Moved gates usage report to binding_view
